requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4

pytest==8.0.2
requests==2.31.0
httpx==0.27.0

scikit-learn==1.4.2
joblib==1.4.2
mlflow==2.12.2
matplotlib==3.8.4